refactor: Move duplicate performance test code into utility class
