+++ src/test/java/com/synopsys/integration/alert/performance/LargeNotificationTest.java
@@ -4,10 +4,8 @@ import static org.junit.jupiter.api.Assertions.fail;
 
 import java.time.Duration;
 import java.time.LocalDateTime;
-import java.time.OffsetDateTime;
 import java.time.format.DateTimeFormatter;
 import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
@@ -28,22 +26,17 @@ import org.springframework.web.context.WebApplicationContext;
 import com.google.gson.Gson;
 import com.synopsys.integration.alert.Application;
 import com.synopsys.integration.alert.api.common.model.ValidationResponseModel;
-import com.synopsys.integration.alert.channel.jira.server.descriptor.JiraServerDescriptor;
 import com.synopsys.integration.alert.channel.jira.server.model.JiraServerGlobalConfigModel;
-import com.synopsys.integration.alert.common.descriptor.ChannelDescriptor;
-import com.synopsys.integration.alert.common.enumeration.FrequencyType;
-import com.synopsys.integration.alert.common.rest.AlertRestConstants;
 import com.synopsys.integration.alert.common.rest.model.FieldValueModel;
 import com.synopsys.integration.alert.configuration.ApplicationConfiguration;
 import com.synopsys.integration.alert.database.DatabaseDataSource;
-import com.synopsys.integration.alert.descriptor.api.BlackDuckProviderKey;
 import com.synopsys.integration.alert.descriptor.api.JiraServerChannelKey;
 import com.synopsys.integration.alert.performance.utility.AlertRequestUtility;
 import com.synopsys.integration.alert.performance.utility.BlackDuckProviderService;
 import com.synopsys.integration.alert.performance.utility.ConfigurationManagerV2;
 import com.synopsys.integration.alert.performance.utility.IntegrationPerformanceTestRunnerV2;
+import com.synopsys.integration.alert.performance.utility.jira.server.JiraServerPerformanceUtility;
 import com.synopsys.integration.alert.test.common.TestProperties;
-import com.synopsys.integration.alert.test.common.TestPropertyKey;
 import com.synopsys.integration.alert.test.common.TestTags;
 import com.synopsys.integration.alert.util.DescriptorMocker;
 import com.synopsys.integration.blackduck.api.generated.view.ProjectVersionView;
@@ -56,9 +49,7 @@ import com.synopsys.integration.exception.IntegrationException;
 @TestPropertySource(locations = "classpath:spring-test.properties")
 @WebAppConfiguration
 class LargeNotificationTest {
-    private static final String PERFORMANCE_JOB_NAME = "Jira Server Performance Job";
     private static final JiraServerChannelKey CHANNEL_KEY = new JiraServerChannelKey();
-    private static final BlackDuckProviderKey PROVIDER_KEY = new BlackDuckProviderKey();
 
     @Autowired
     private WebApplicationContext webApplicationContext;
@@ -70,6 +61,7 @@ class LargeNotificationTest {
     private AlertRequestUtility alertRequestUtility;
     private BlackDuckProviderService blackDuckProviderService;
     private ConfigurationManagerV2 configurationManager;
+    private JiraServerPerformanceUtility jiraServerPerformanceUtility;
 
     @BeforeEach
     public void init() {
@@ -81,6 +73,7 @@ class LargeNotificationTest {
             blackDuckProviderService.getBlackDuckProviderKey(),
             CHANNEL_KEY.getUniversalKey()
         );
+        jiraServerPerformanceUtility = new JiraServerPerformanceUtility(alertRequestUtility, configurationManager);
     }
 
     @Test
@@ -105,7 +98,7 @@ class LargeNotificationTest {
     void largeNotificationTest() throws IntegrationException, InterruptedException {
         LocalDateTime startingTime = LocalDateTime.now();
         TestProperties testProperties = new TestProperties();
-        JiraServerGlobalConfigModel jiraServerGlobalConfigModel = createGlobalConfigModel(testProperties);
+        JiraServerGlobalConfigModel jiraServerGlobalConfigModel = jiraServerPerformanceUtility.createGlobalConfigModel(testProperties);
 
         // Create Black Duck Global Provider configuration
         LocalDateTime startingNotificationSearchDateTime = LocalDateTime.now();
@@ -113,19 +106,19 @@ class LargeNotificationTest {
         logTimeElapsedWithMessage("Triggering the Black Duck notification took %s", startingNotificationSearchDateTime, LocalDateTime.now());
 
         // Create Jira Server global config
-        ValidationResponseModel installPluginResponse = installPlugin(jiraServerGlobalConfigModel);
+        ValidationResponseModel installPluginResponse = jiraServerPerformanceUtility.installPlugin(jiraServerGlobalConfigModel);
         if (installPluginResponse.hasErrors()) {
             fail("Unable to install the Alert plugin for Jira Server. Exiting test...");
         }
 
-        Optional<JiraServerGlobalConfigModel> globalConfiguration = createGlobalConfiguration(jiraServerGlobalConfigModel);
+        Optional<JiraServerGlobalConfigModel> globalConfiguration = jiraServerPerformanceUtility.createGlobalConfiguration(jiraServerGlobalConfigModel);
         if (globalConfiguration.isEmpty()) {
             fail("Global configuration missing.");
         }
 
         // Create distribution job
         // TODO: Look into adding this feature in IntegrationPerformanceTestRunnerV2
-        Map<String, FieldValueModel> channelFieldsMap = createChannelFieldsMap(testProperties, globalConfiguration.get().getId());
+        Map<String, FieldValueModel> channelFieldsMap = jiraServerPerformanceUtility.createChannelFieldsMap(testProperties, globalConfiguration.get().getId());
 
         // Create N number of blackduck projects
         List<ProjectVersionWrapper> projectVersionWrappers = createBlackDuckProjects(10);
@@ -161,98 +154,6 @@ class LargeNotificationTest {
         logTimeElapsedWithMessage("Triggering the Black Duck notification took %s", startingNotificationTriggerDateTime, LocalDateTime.now());
     }
 
-    //TODO: put into utility class
-    private Map<String, FieldValueModel> createChannelFieldsMap(TestProperties testProperties, String globalConfigId) {
-        Map<String, FieldValueModel> channelFieldsMap = new HashMap<>();
-        channelFieldsMap.put(ChannelDescriptor.KEY_ENABLED, createFieldValueModel("true"));
-        channelFieldsMap.put(ChannelDescriptor.KEY_CHANNEL_NAME, createFieldValueModel(CHANNEL_KEY.getUniversalKey()));
-        channelFieldsMap.put(ChannelDescriptor.KEY_NAME, createFieldValueModel(PERFORMANCE_JOB_NAME));
-        channelFieldsMap.put(ChannelDescriptor.KEY_FREQUENCY, createFieldValueModel(FrequencyType.REAL_TIME.name()));
-        channelFieldsMap.put(ChannelDescriptor.KEY_PROVIDER_TYPE, createFieldValueModel(PROVIDER_KEY.getUniversalKey()));
-        channelFieldsMap.put(ChannelDescriptor.KEY_CHANNEL_GLOBAL_CONFIG_ID, createFieldValueModel(globalConfigId));
-
-        channelFieldsMap
-            .put(JiraServerDescriptor.KEY_ADD_COMMENTS, createFieldValueModel(testProperties.getOptionalProperty(TestPropertyKey.TEST_JIRA_SERVER_ADD_COMMENTS).orElse("true")));
-        channelFieldsMap
-            .put(JiraServerDescriptor.KEY_ISSUE_CREATOR, createFieldValueModel(testProperties.getOptionalProperty(TestPropertyKey.TEST_JIRA_SERVER_ISSUE_CREATOR).orElse("")));
-        channelFieldsMap
-            .put(JiraServerDescriptor.KEY_JIRA_PROJECT_NAME, createFieldValueModel(testProperties.getProperty(TestPropertyKey.TEST_JIRA_SERVER_PROJECT_NAME)));
-        channelFieldsMap
-            .put(JiraServerDescriptor.KEY_ISSUE_TYPE, createFieldValueModel(testProperties.getOptionalProperty(TestPropertyKey.TEST_JIRA_SERVER_ISSUE_TYPE).orElse("Task")));
-        channelFieldsMap
-            .put(
-                JiraServerDescriptor.KEY_RESOLVE_WORKFLOW_TRANSITION,
-                createFieldValueModel(testProperties.getOptionalProperty(TestPropertyKey.TEST_JIRA_SERVER_RESOLVE_TRANSITION).orElse(""))
-            );
-        channelFieldsMap
-            .put(
-                JiraServerDescriptor.KEY_OPEN_WORKFLOW_TRANSITION,
-                createFieldValueModel(testProperties.getOptionalProperty(TestPropertyKey.TEST_JIRA_SERVER_REOPEN_TRANSITION).orElse(""))
-            );
-
-        return channelFieldsMap;
-    }
-
-    //TODO: put into utility class
-    private FieldValueModel createFieldValueModel(String value) {
-        return new FieldValueModel(List.of(value), true);
-    }
-
-    //TODO: put into utility class
-    private Optional<JiraServerGlobalConfigModel> createGlobalConfiguration(JiraServerGlobalConfigModel jiraServerGlobalConfigModel) {
-        String apiConfigurationPath = AlertRestConstants.JIRA_SERVER_CONFIGURATION_PATH;
-        if (null != jiraServerGlobalConfigModel) {
-            LocalDateTime startingTime = LocalDateTime.now();
-            String descriptorName = CHANNEL_KEY.getUniversalKey();
-            Optional<JiraServerGlobalConfigModel> globalConfigModel = configurationManager
-                .createGlobalConfiguration(apiConfigurationPath, jiraServerGlobalConfigModel, JiraServerGlobalConfigModel.class);
-            String globalConfigMessage = String.format("Creating the global Configuration for %s jobs took", descriptorName);
-            logTimeElapsedWithMessage(globalConfigMessage + " %s", startingTime, LocalDateTime.now());
-            return globalConfigModel;
-        }
-        return Optional.empty();
-    }
-
-    //TODO: put into utility class
-    private JiraServerGlobalConfigModel createGlobalConfigModel(TestProperties testProperties) {
-        String createdAt = OffsetDateTime.now().toString();
-        String url = testProperties.getProperty(TestPropertyKey.TEST_JIRA_SERVER_URL);
-        String userName = testProperties.getProperty(TestPropertyKey.TEST_JIRA_SERVER_USERNAME);
-        String password = testProperties.getProperty(TestPropertyKey.TEST_JIRA_SERVER_PASSWORD);
-        Boolean disablePluginCheck = testProperties.getOptionalProperty(TestPropertyKey.TEST_JIRA_SERVER_DISABLE_PLUGIN_CHECK)
-            .map(Boolean::valueOf)
-            .orElse(Boolean.FALSE);
-
-        return new JiraServerGlobalConfigModel(
-            null,
-            AlertRestConstants.DEFAULT_CONFIGURATION_NAME,
-            createdAt,
-            createdAt,
-            url,
-            userName,
-            password,
-            Boolean.FALSE,
-            disablePluginCheck
-        );
-    }
-
-    //TODO: put into utility class
-    private ValidationResponseModel installPlugin(JiraServerGlobalConfigModel jiraServerGlobalConfigModel) {
-        try {
-            String requestBody = gson.toJson(jiraServerGlobalConfigModel);
-            String installPluginResponseString = alertRequestUtility.executePostRequest(
-                String.format("%s/install-plugin", AlertRestConstants.JIRA_SERVER_CONFIGURATION_PATH),
-                requestBody,
-                "Installing the plugin failed."
-            );
-            return gson.fromJson(installPluginResponseString, ValidationResponseModel.class);
-        } catch (IntegrationException e) {
-            logger.error("Unexpected error occurred while installing the plugin.", e);
-            return ValidationResponseModel.generalError(e.getMessage());
-        }
-    }
-
-    //TODO: put into utility class
     public void logTimeElapsedWithMessage(String messageFormat, LocalDateTime start, LocalDateTime end) {
         //TODO log timing to a file
         Duration duration = Duration.between(start, end);
